my_env/requirements.txt
openai>=1.0.0
PyYAML>=6.0
fastapi>=0.110.0
uvicorn>=0.27.0
openenv-core>=0.2.0
uv>=0.5.0
